requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements/base.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
boto3==1.34.84
    # via -r requirements/base.in
botocore==1.34.84
    # via
    #   boto3
    #   s3transfer
certifi==2024.2.2
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.3.2
    # via requests
dj-database-url==2.1.0
    # via -r requirements/base.in
django==4.2.11
    # via
    #   -r requirements/base.in
    #   dj-database-url
    #   django-cors-headers
    #   django-storages
    #   djangorestframework
    #   drf-yasg
    #   sentry-sdk
django-cors-headers==4.3.1
    # via -r requirements/base.in
django-environ==0.11.2
    # via -r requirements/base.in
django-storages==1.14.2
    # via -r requirements/base.in
djangorestframework==3.15.1
    # via
    #   -r requirements/base.in
    #   drf-yasg
drf-yasg==1.21.7
    # via -r requirements/base.in
idna==3.7
    # via requests
inflection==0.5.1
    # via drf-yasg
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
packaging==24.0
    # via drf-yasg
pillow==10.3.0
    # via -r requirements/base.in
psycopg2-binary==2.9.9
    # via -r requirements/base.in
python-dateutil==2.9.0.post0
    # via botocore
pytz==2024.1
    # via drf-yasg
pyyaml==6.0.1
    # via drf-yasg
requests==2.32.3
    # via -r requirements/base.in
s3transfer==0.10.1
    # via boto3
sentry-sdk[django]==1.45.0
    # via -r requirements/base.in
six==1.16.0
    # via python-dateutil
sqlparse==0.5.0
    # via django
typing-extensions==4.11.0
    # via
    #   asgiref
    #   dj-database-url
uritemplate==4.1.1
    # via drf-yasg
urllib3==1.26.18
    # via
    #   -r requirements/base.in
    #   botocore
    #   requests
    #   sentry-sdk
whitenoise==6.7.0
    # via -r requirements/base.in

####
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0